Added some data checks for the Use cases

+++ src/controllers/user.controller.js
@@ -48,19 +48,21 @@ let userController = {
                     let queryValue2 = queryField[1][1]
 
                     let myBool
-                    if(queryField1 === 'isActive' && (queryValue1 === 'true' || queryValue1 === 'false')) {
-                        myBool = (queryValue1 === 'true')
-                        queryValue1 = myBool
-                    } else if(queryField1 === 'id') {
+                    // if(queryField1 === 'isActive' && (queryValue1 === 'true' || queryValue1 === 'false')) {
+                    //     myBool = (queryValue1 === 'true')
+                    //     queryValue1 = myBool
+                    // } 
+                    if(queryField1 === 'id' || queryField1 === 'isActive') {
                         myBool = parseInt(queryValue1)
                         queryValue1 = myBool
                     }
 
                     let myBool2
-                    if(queryField2 === 'isActive' && (queryValue2 === 'true' || queryValue2 === 'false')) {
-                        myBool2 = (queryValue2 === 'true')
-                        queryValue2 = myBool2
-                    } else if(queryField2 === 'id') {
+                    // if(queryField2 === 'isActive' && (queryValue2 === 'true' || queryValue2 === 'false')) {
+                    //     myBool2 = (queryValue2 === 'true')
+                    //     queryValue2 = myBool2
+                    // }  
+                    if(queryField2 === 'id' || queryField2 === 'isActive') {
                         myBool2 = parseInt(queryValue2)
                         queryValue2 = myBool2
                     }
@@ -86,10 +88,11 @@ let userController = {
                     let queryValueLength1 = queryField[0][1]
 
                     let myBool1
-                    if(queryFieldLength1 === 'isActive' && (queryValueLength1 === 'true' || queryValueLength1 === 'false')) {
-                        myBool1 = (queryValueLength1 === 'true')
-                        queryValueLength1 = myBool1
-                    } else if(queryFieldLength1 === 'id') {
+                    // if(queryFieldLength1 === 'isActive' && (queryValueLength1 === 'true' || queryValueLength1 === 'false')) {
+                    //     myBool1 = (queryValueLength1 === 'true')
+                    //     queryValueLength1 = myBool1
+                    // } 
+                    if(queryFieldLength1 === 'id' || queryFieldLength1 === 'isActive') {
                         myBool1 = parseInt(queryValueLength1)
                         queryValueLength1 = myBool1
                     }
@@ -157,7 +160,9 @@ let userController = {
         let myUserId = userId.substring(1)
         const numberUserId = parseInt(myUserId)
 
-        userService.deleteUser(numberUserId, (error, success) => {
+        const userIdFromToken = req.userId
+
+        userService.deleteUser(numberUserId, userIdFromToken, (error, success) => {
             if (error) {
                 return next({
                     status: error.status,
@@ -181,7 +186,10 @@ let userController = {
         const numberUserId = parseInt(myUserId)
         const user = req.body
 
-        userService.changeUser(user, numberUserId, (error, success) => {
+        const userIdFromToken = req.userId
+
+
+        userService.changeUser(user, numberUserId, userIdFromToken, (error, success) => {
             if (error) {
                 return next({
                     status: error.status,
